libsmartctl: refactored some stuff out of ataprint.cpp and smartctl.cpp so libsmartctl no longer needs dependent on ataprint.cpp or require functions copied over from smartctl.cpp
